[franka_hand] Add RS camera TF tree in xacro
The camera TF tree, taken from the official Realsense D435 description,
is connected to the Franka hand tree through the camera calibration
result: the TF between the hand and optical camera frames.
The rest of the camera frames are connected to the optical camera frame
to match the TFs from the original camera description.